packages/services/agent-looper/python/requirements.txt
# Agent Looper Requirements for GraphWiz-XR

# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1

# gRPC (using compatible versions)
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf==4.25.1

# Async support
asyncio>=3.4.3

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1
